requirements.txt
Django==4.2.29
django-allauth==65.14.3
Jinja2==3.1.6
python-dotenv==1.2.1
PyMySQL==1.1.2
whitenoise==6.11.0
Pillow==11.3.0
requests==2.32.5
PyJWT==2.12.1
cryptography==46.0.5
urllib3==1.26.20
pypdf==6.9.0
pdfplumber==0.11.8
